+++ packages/ui/src/__tests__/molecules/InputGroup.test.tsx
@@ -17,7 +17,8 @@ describe('InputGroupText', () => {
 
     expect(label).toBeInTheDocument()
     expect(input).toBeInTheDocument()
-    expect(label).toHaveAttribute('for', input.id)
+    expect(label).toHaveAttribute('for', 'testid')
+    expect(input).toHaveAttribute('id', 'testid')
     expect(input).toHaveAttribute('name', 'testname')
   })