+++ packages/ui/src/components/atoms/index.ts
@@ -1,8 +1,9 @@
 import Input from './Input'
 import Label from './Label'
-import { Icon } from './Icon'
+import Icon  from './Icon'
 import { Button } from './Button'
 import HamburgerMenu from './HamburgerMenu'
+import ValidationMessage from './ValidationMessage'
 import Radio from './Radio'
 
-export { Input, Label, Button, HamburgerMenu, Icon, Radio }
+export { Input, Label, Button, HamburgerMenu, Icon, ValidationMessage, Radio }