+++ packages/ui/src/components/molecules/index.ts
@@ -1,3 +1,4 @@
 import { BackButton } from "./BackButton";
+import  InputGroup  from "./InputGroup";
 
-export { BackButton }
+export { BackButton , InputGroup}