+++ packages/ui/src/stories/atoms/Icon.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Icon } from '../../components/atoms/Icon'
+import Icon from '../../components/atoms/Icon'
 import { font } from '../../styles/font'
 
 // More on how to set up stories at: https://storybook.js.org/docs/writing-stories#default-export