+++ sso/src/controllers/auth/registerController.ts
@@ -6,12 +6,13 @@ import { generateId } from '../../utils/cuidGenerator'
 
 export const registerController: Middleware = async (ctx: Context) => {
   const { dni, email, password, itineraryId }: UserRegister = ctx.request.body
+  const dniToUpperCase = dni.toUpperCase()
   const hashedPassword = await hashPassword(password)
   const id = generateId()
 
   const query = {
     text: 'INSERT INTO "user"(id, dni, email, password, user_meta, itinerary_id) VALUES($1, $2, $3, $4, $5, $6)',
-    values: [id, dni, email, hashedPassword, '{}', itineraryId],
+    values: [id, dniToUpperCase, email, hashedPassword, '{}', itineraryId],
   }
 
   await client.query(query)